refactor:#68 useEffect 커스텀훅으로 분리
